Better style for sidebar bookmarks
