+++ app/components/account/action/add/batch-account-create.html
@@ -1,5 +1,5 @@
 <bl-complex-form [formGroup]="form" [submit]="submit" [containerRef]="sidebarRef">
-    <bl-form-page main-form-page title="Create batch account" [formGroup]="form">
+    <bl-form-page main-form-page [title]="title" [formGroup]="form">
         <bl-form-section title="General info" subtitle="Basic information about the batch account">
             <div class="form-element">
                 <mat-form-field>
