+++ src/@batch-flask/ui/sidebar/index.ts
@@ -3,4 +3,4 @@ export * from "./sidebar-content";
 export * from "./sidebar-page";
 export * from "./sidebar-ref";
 export * from "./sidebar-bookmarks";
-export * from "./sidebar";
+export * from "./sidebar.module";
