+++ src/@batch-flask/ui/sidebar/sidebar-bookmarks/index.ts
@@ -1,0 +1,1 @@
+export * from "./sidebar-bookmarks.component";
