+++ src/@batch-flask/ui/sidebar/sidebar-bookmarks/sidebar-bookmarks.scss
@@ -2,38 +2,46 @@
 
 bl-sidebar-bookmarks {
     [bl-dropdown-content] {
-        width: 200px;
+        min-width: 200px;
+        max-width: 400px;
     }
 
     .no-bookmarks {
         padding: 10px 5px;
     }
 
-    .bookmark {
+    .dropdown-item.bookmark {
         position: relative;
-        min-width: 100px;
+        min-width: 250px;
         cursor: pointer;
-        padding: 10px 5px;
+        padding: 0;
         display: flex;
-        align-items: center;
-        justify-content: center;
+        align-items: stretch;
+        justify-content: stretch;
 
-        span {
-            flex: 1 0 auto;
-            margin-right: 5px;
+        &:hover {
+            i.fa {
+                color: $primary-color-dark;
+            }
+        }
+        > span {
+            flex: 1 1 auto;
             text-align: left;
+            min-width: 20px;
+            text-overflow: ellipsis;
+            overflow-x: hidden;
+            border-right: 1px solid $border-color;
+            padding: 5px 10px;
         }
 
-        i.fa {
-            flex: 0 0 20px;
+        > i.fa {
+            flex: 0 0 30px;
             background-color: white;
             color: $primary-color-dark;
-            border-radius: 10px;
-            width: 20px;
-            height: 20px;
-            text-align: center;
-            vertical-align: middle;
-            line-height: 20px;
+            display: flex;
+            align-items: center;
+            justify-content: center;
+
             cursor: pointer;
 
             &:hover {
