+++ src/@batch-flask/ui/sidebar/sidebar.module.ts
@@ -1,4 +1,4 @@
-import { ModuleWithProviders, NgModule } from "@angular/core";
+import { NgModule } from "@angular/core";
 import { FormsModule } from "@angular/forms";
 import { BrowserModule } from "@angular/platform-browser";
 import { MaterialModule } from "@batch-flask/core";
@@ -9,19 +9,19 @@ import { SidebarContentComponent } from "./sidebar-content";
 import { SidebarManager } from "./sidebar-manager";
 import { SidebarPageComponent } from "./sidebar-page";
 
+const publicComponents = [
+    SidebarContentComponent,
+    SidebarPageComponent,
+    SidebarBookmarksComponent,
+];
+const privateComponents = [];
+
 @NgModule({
-    declarations: [
-        SidebarContentComponent,
-        SidebarPageComponent,
-        SidebarBookmarksComponent,
-    ],
+    declarations: [...publicComponents, ...privateComponents],
     entryComponents: [
         SidebarPageComponent,
     ],
-    exports: [
-        SidebarContentComponent,
-        SidebarPageComponent,
-        SidebarBookmarksComponent],
+    exports: [...publicComponents],
     imports: [
         BrowserModule,
         FormsModule,
@@ -33,10 +33,4 @@ import { SidebarPageComponent } from "./sidebar-page";
     ],
 })
 export class SidebarModule {
-    public static forRoot(): ModuleWithProviders {
-        return {
-            ngModule: SidebarModule,
-            providers: [],
-        };
-    }
 }
